Refactor constants and helper methods
