+++ control-statistics/src/main/java/fi/nls/oskari/control/statistics/GetIndicatorMetadataHandler.java
@@ -10,13 +10,8 @@ import fi.nls.oskari.control.statistics.data.*;
 import fi.nls.oskari.control.statistics.plugins.*;
 import fi.nls.oskari.domain.User;
 import fi.nls.oskari.util.ResponseHelper;
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
 
 /**
  * This interface gives the relevant information for one indicator to the frontend.
@@ -31,8 +26,6 @@ import java.util.Map;
 @OskariActionRoute("GetIndicatorMetadata")
 public class GetIndicatorMetadataHandler extends ActionHandler {
     private final static String CACHE_PREFIX = "oskari_get_indicator_metadata_handler_";
-    private final static String PARAM_PLUGIN_ID = "datasource";
-    private final static String PARAM_INDICATOR_ID = "indicator";
 
     /**
      * For now, this uses pretty much static global store for the plugins.
@@ -42,8 +35,8 @@ public class GetIndicatorMetadataHandler extends ActionHandler {
 
     @Override
     public void handleAction(ActionParameters ap) throws ActionException {
-        final long pluginId = ap.getRequiredParamInt(PARAM_PLUGIN_ID);
-        final String indicatorId = ap.getRequiredParam(PARAM_INDICATOR_ID);
+        final long pluginId = ap.getRequiredParamInt(StatisticsHelper.PARAM_DATASOURCE_ID);
+        final String indicatorId = ap.getRequiredParam(StatisticsHelper.PARAM_INDICATOR_ID);
         JSONObject response = getIndicatorMetadataJSON(ap.getUser(), pluginId, indicatorId);
         ResponseHelper.writeResponse(ap, response);
     }
@@ -74,7 +67,7 @@ public class GetIndicatorMetadataHandler extends ActionHandler {
             throw new ActionParamsException("No such indicator: " + indicatorId + " on datasource: " + pluginId);
         }
         try {
-            JSONObject indicatorMetadata = toJSON(indicator);
+            JSONObject indicatorMetadata = StatisticsHelper.toJSON(indicator);
             // Note that there is an another layer of caches in the plugins doing the web queries.
             // Two layers are necessary, because deserialization and conversion to the internal data model
             // is pretty heavy operation.
@@ -85,53 +78,6 @@ public class GetIndicatorMetadataHandler extends ActionHandler {
         } catch (JSONException e) {
             throw new ActionException("Something went wrong in getting indicator metadata.", e);
         }
-    }
-    
-    public static JSONObject toJSON(StatisticalIndicator indicator) throws JSONException {
-        JSONObject pluginIndicatorJSON = new JSONObject();
-        Map<String, String> name = indicator.getName();
-        Map<String, String> description = indicator.getDescription();
-        Map<String, String> source = indicator.getSource();
-        List<StatisticalIndicatorLayer> layers = indicator.getLayers();
-        StatisticalIndicatorDataModel selectors = indicator.getDataModel();
-
-        pluginIndicatorJSON.put("id", indicator.getId());
-        pluginIndicatorJSON.put("name", name);
-        pluginIndicatorJSON.put("description", description);
-        pluginIndicatorJSON.put("source", source);
-        pluginIndicatorJSON.put("public", indicator.isPublic());
-        pluginIndicatorJSON.put("regionsets", toJSON(layers));
-        pluginIndicatorJSON.put("selectors", toJSON(selectors));
-        return pluginIndicatorJSON;
-    }
-
-    public static JSONArray toJSON(StatisticalIndicatorDataModel selectors) throws JSONException {
-        JSONArray selectorsJSON = new JSONArray();
-        for (StatisticalIndicatorDataDimension selector : selectors.getDimensions()) {
-            JSONObject selectorJSON = new JSONObject();
-            selectorJSON.put("id", selector.getId());
-            selectorJSON.put("name", selector.getName());
-            selectorJSON.put("allowedValues", toJSON(selector.getAllowedValues()));
-            // Note: Values are not given here, they are null anyhow in this phase.
-            selectorsJSON.put(selectorJSON);
-        }
-        return selectorsJSON;
-    }
-
-    private static JSONArray toJSON(Collection<IdNamePair> stringCollection) {
-        JSONArray stringArray = new JSONArray();
-        for (IdNamePair value : stringCollection) {
-            stringArray.put(value.getValueForJson());
-        }
-        return stringArray;
-    }
-
-    public static JSONArray toJSON(List<StatisticalIndicatorLayer> layers) throws JSONException {
-        JSONArray layersJSON = new JSONArray();
-        for (StatisticalIndicatorLayer layer: layers) {
-            layersJSON.put(layer.getOskariLayerId());
-        }
-        return layersJSON;
     }
 
 }
